Update rand in the stdlib tests, and remove the getrandom feature from it
